既存ファイルをcreate-viteで生成されたファイルで置き換え
