+++ electron/electron-env.d.ts
@@ -23,5 +23,5 @@ declare namespace NodeJS {
 
 // Used in Renderer process, expose in `preload.ts`
 interface Window {
-  ipcRenderer: import('electron').IpcRenderer
+  osc: import('./preload').OscApi;
 }
