+++ src/main.ts
@@ -1,4 +1,5 @@
 import { createApp } from 'vue';
 import App from './App.vue';
+import './assets/style.css';
 
 createApp(App).mount('#app');
